organize tests and add one test for invariant locale compare as exaple
